% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation - ransampler.R
\docType{package}
\name{Genotools}
\alias{Genotools}
\title{Ransampler - A tool for doing random (condition based, grouping and no-sharing) samples of a population}
\description{
Ransampler - A tool for doing random (condition based, grouping and no-sharing) samples of a population
}
