% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - ransampler 2.R
\name{ransampler}
\alias{ransampler}
\title{Random, condition based samplig of a set of individuals}
\usage{
ransampler(
  table,
  ofEach,
  except,
  nOfEach = 1,
  noShareWithin = c(),
  priBy,
  useDuplis = F,
  identifier = "",
  returnCombTable = F,
  runs = 1,
  reshuffle_combinationTable = F
)
}
\arguments{
\item{ofEach}{The attributes we want to randomise over, as a vector of column names, e.g: ofEach=c("sex","age","tank")}

\item{noShareWithin}{One or more sets of paremeters, where unique combinations can't be shared. For example, if we want no more than a single individual from any given family, within a tank, a set would be c("family","tank"), meaning that no individuals can have the same of both family and tank. Supplied within a list, so: list(c("family","tank")). Add more sets within the list if needed, for example: list(c("tank","father"),c("tank","mother)))}

\item{priBy}{If some individuals are to be prioritized over others, specify the name of the column containing prioritization info. This must ba number, and lower numbers are prioritized. (E.g, individuals with "1" are prioritized over individuals iwth "2")}

\item{useDuplis}{If all individuals that are selected within each combination may be be used, or just one of them.}

\item{dataframe}{The dataframe to get the individuals from. Each row must be a single individual. Columns are attributes.}

\item{N_ofEach}{How many individuals within each combination of the "ofEach" parameters that should be selected}
}
\description{
Takes a dataframe of individuals (rows) with certain attributes (columns), and samples a given number of individuals from all possible combinations of a given set of paramters
For example, can sample one individual of ech combination of sex and age
Also allows for some control, like, "no individuals sharing the same treatment and family"
}
