% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadesc-salmon.R
\docType{data}
\name{table_salmon}
\alias{table_salmon}
\title{salmon study population}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4507 rows and 12 columns.
}
\usage{
data(table_salmon)
}
\description{
example data for ransampler
individual fish data, with length, weight, sex, vgll3 genotype, population, and family information
}
\keyword{datasets}
