% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadesc-salmon-small.R
\docType{data}
\name{table_salmon_small}
\alias{table_salmon_small}
\title{salmon study population}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1625 rows and 11 columns.
}
\usage{
data(table_salmon_small)
}
\description{
example data for ransampler
individual fish data, with length, weight, sex, vgll3 genotype, population, and family information
Smaller dataset with fewer individuals
}
\keyword{datasets}
